Remove the use of JNA CallbackThreadInitializers.
This was originally added because it seemed JNA was creating a new
heavyweight Java thread each time a callback was invoked from native
code. 

The problem is this thread NEVER gets detached and if you, for example,
stop, play, stop, play media repeatedly you will leak objects. 

In actual fact, JNA does not create heavyweight Java threads, they are
lightweight proxies to real threads. So while it looks like many threads
are being created causing a perceived performance problem, it is not
actually the case.