Attempt to transparently handle local filenames with Unicode characters.
When specifying an MRL, if that MRL appears to be a local file (it does
not have a URL scheme like http, file or whatever), and if that MRL also
contains at least one Unicode character, then the MRL is converted to a
file URL and the Unicode characters in the filename are percent-encoded.

Without this, vlc will be unable to play the MRL since vlc uses a native
ASCII API that requires Unicode characters to be encoded in this way.

Any MRL that looks like a URL will not be encoded by this method, only
local filenames will be encoded.