Small improvements to the adaptive layout test.
The example as well as minimizing the video surface size to 1x1, the
minimum size that will still work, it now uses a CardLayout to switch
away from the video view to a blank container.

If you don't switch to a blank container, you will still see a 1x1 pixel
video playing.