Support embedding of video surface when using the direct audio player.
There is no reason why a direct audio player can not also play video in
an associated video surface like an embedded media player.

In this way, it is possible to play video, and do analysis or whatever
on the raw audio samples before playing them (by JavaSound or some other
Java API for playing sound).