Provide a system property to disable automatic LibX initialisation.
This affects Linux only.

It is not clear whether or not LibX.initialise() is always needed
on Linux - some versions of VLC will complain if it is not invoked.

If it is invoked, it can conflict with use of the JVM splashscreen
options (and lead to a JVM crash).

Setting the VLCJ_INITX system property to "no" will prevent the
automatic invocation.

e.g. java -DVLCJ_INITX=no ...