Support new native audio events.
It is now possible, with LibVLC 3.0.0+ to get audio mute/volume
changes as well as corking/uncorking.

This means that it should now be possible to reliably set/get
initial volume and mute status for a media player (from the OS).

Changing volume/mute with OS controls will now be received as
events in a vlcj application, and changing volume/mute in a vlcj
application will be reflected in the OS controls.