Raise new event when the media has actually started playing
This differs from the playing event, and the videoOutput event, in
that it fires once after the media has started and the first native
position-changed event has been received.

This is required for things like logo and marquee (and other things)
that can not be successfully set in the playing or videoOutput
events (they are too early).

The event will be fired once again if the media is stopped (or is
finished) and played again.

It will not be fired on a pause/play or position/time change.