First stage refactoring to stop using a single global instance of LibVLC
The major part of this change was to refactor MediaPlayerFactory.

The factory now uses a NativeDiscovery implementation to automatically
find the native library (previously this was left up to the application
developer).

The factory implementation has been refactored to reduce the size and
complexity of this large class. The various factory methods that create
things like media players, video surfaces and so on have had their
implementations moved to private "service" classes. Unfortunately this
results in significant backwards-compatibility API breakage.

A further consequence of this change is that LibVlcFactory has been
removed as it is no longer needed.