First pass at major refactoring of MediaPlayer and MediaPlayerFactory
The biggest changes are factoring the factory to its own package,
and splitting the MediaPlayerFactory and media player implementation
mega-classes into smaller delegating components. Consequently there
is likely to be significant API breakage for client applications.

There are also a lot of internal changes and reorganisations to make
the code smaller and easier to maintain going forward - this should
not impact client applications - and the actual implementations inside
components is generally unchanged.

Some problematic or duplicative API has been removed - impact on
client application in this regard is expected to be minimal.

This pass has ignored Javadoc updates which will be addressed later.

A bunch of developer comments and FIXME remain, which will be
resolved in due course.