First pass of refactoring Media into a top-level concept
These extensive changes unpick the existing media and sub-item
implementation from the base media player and introduces a Media
component as a top-level component in its own right.

Event handling for media and sub-items was removed from the
media player and instead is found in the new Media component.

Parsing and meta-data operations have been moved from the base
media palyer to the new Media component.

Sub-item implementation has improved drastically and is now much
more simple and more robust, taking advantage of a native media
list player to handle sub-items and removing a lot of brittle
hard-to-grok sub-item code from the base media player.

Applications may no longer call methods like playMedia(mrl, options)
on the media player object - instead a Media instance must be
created via the MediaPlayerFactory, options may then be added,
and this media object is then "set" on the media player and played
via media player controls.

There were a LOT of knock-on changes required.

Tests need to be revisited to bring up to spec with the new API.

A Javadoc pass is required.