On OSX, try the reflection method to get a window handle
On JDK 1.6, Native.getComponentID should still work, so try that
first - if this throws an exception use the brittle reflection
code to dig deep into the OSX private implementation classes to
try and pick out a usable window handle.

Note that even if this works, the window associated with the
returned handle does not seem to fire the resize events that VLC
needs. Consequently it is not possible to dynamically resize the
video output (at the moment, perhaps forever). If you stop then
start the media player the new size *is* picked up.