Fix hue video adjustment
This was previously an integer type with a range from 0 to 360. The
native implementation as changed at some point to make this a float
type with a range fro -180 to 180.

Since it was no longer an int type, and vlcj was invoking the int
API, it had no effect.

Changing to use the float API for video adjustment fixes the issue
with hue.