Fix JVM crashes when releasing certain structures
The crashes were occurring because the native API binding was
using a Structure reference as the pointer to release - when the
release native call returns, JNA will try and automatically
synchronise the structure, but in this case the native memory is
gone -> crash.

The solution is to use a Pointer, or PointerType, on the native
API binding since this will not be synchronised.

The old implementation used setAutoSynch(false) on the structure,
which worked, but was unclear.