Improvements to the Native Discovery component
Hopefully this has simplified things and made it easier to extend
the functionality for the most common use-cases.

Some standard implementations are provided for all supported
operating systems and these should work in most cases.

A Service Loader interface provides an easy extension point for
client applications if so desired. Specifically it is possible to
augment the standard directory search locations simply by adding a
bespoke implementation of a DiscoveryDirectoryProvider to the
run-time application classpath (and corresponding META-INF/services
file). A simple priority system is used to define the order in
which the directories are searched.

Client applications are still free to fully implement their own
replacement discovery strategies and it will all still work as
before (though with some API changes).

The key goals of these changes are first to make the default
behaviour just work out-of-the-box, and second for the likely most
common extension use case to make it easy to add more search
directories (and this can be done without interacting directly
with the native discovery mechanism).

These changs have been tested on Linux, OSX (latest), and the only
Windows environment I had access to (not Windows 10, although it
should be fine).