Add a component-spec mini-builder for media player components
A full-blown builder is available for all media player components,
but using this builder prevents a client application from using
one of the main advantages of the media player components - i.e.
that it is trivially easy to add event handlers by dynamically
sub-classing the component.

With the builder, this can't work because the builder is creating
the new instance.

The builder is still useful, but another solution is needed.

To dynamically sub-class the components it is necessary to use
"new" to create the component class - this can be onerous and ugly
 now because the components are now configured via a list of
constructor params, and a client application may end up having to
pass a parameter list containing e.g. four out of 5 nulls.

That is not ideal.

The solution therefore is to implement a very simple mini-builder
pattern for the constructor parameters.

This gives a decent compromise between the full-blown builder and
a long null argument list when using "new".

An application is still free to use "new" and the constructor
parameter list if necessary.