Improvements to the "adaptive" full-screen strategy
Rename public API class to be less verbose.

Provide a way to set/use a "default" or "other" strategy if the
runtime OS check is unknown (this should never happen, but the
default provides a safe fallback).

The implementation picks the best available from the strategies
provided by vlcj. These can all be overridden in a sub-class if a
client application has its own strategies (a client can also use
zero or more of the default implementatons provided for each OS if
this is useful).

In sub-classes, renamed the template methods beforeEnterFullScreen
to onBeforeEnterFullScreen (same for exit).

This change also removes reliance on RuntimeType enum and the
associated method in RuntimeUtil.