Add a native full-screen strategy implementation for OSX
This implementation uses Apple's EAWT classes.

In order to continue to be able to build vlcj on Linux and Windows
stubs for those classes are used (obviously the EAWT classes are
only available on OSX). The stubs are in a separate project with
a scope of "provided", so are needed only at compile-time and should
NOT be distributed with applications.