Only create native callback and CTI if there is a listener
Previously the callback instance and the corresponding callback
thread initialiser (CTI) were created (but not registered) even if
there was no active listener.

This was unnecessary and so has been changed to only do that when
needed.

Similarly, when there are no longer any listeners, and all the
native events have been unregistered, the CTI is explicitly
detached from the callback.