Cleaned up the Media Resource Locator encoding heuristic
Renamed the encode method to "toLocalFileUri" to be explicitly clear
about the intent and scope of the method.

Removed the test for an MRL representing an absolute file path. This
test was redundant, since if the MRL was a local file (and it must
be either local or absolute) there would be no URISyntaxException
thrown. The catch block therefore can be considered an else clause
so there is no need to check the file path before deciding to encode
the MRL or not.