Changes to URI handling generally for URI's passed to native API
URI encoding is required in a couple of places - MRL's when playing
media, and URI's when adding input slaves (i.e. adding subtitles
or additional audio tracks).

This library has always tried to hide some complexities regarding
MRL's, enabling client applications to easily play media without
having to be concerned with some of the fine detail when using
URI's.

In particular, playing local files with non-ASCII characters in
their names can cause problems on some operating systems (mainly
Windows is where the problems are), so we have to detect those
file names and convert them into proper URL's, with a proper
"file:///" prefix and percent encoding of the non-ASCII characters
so we end up with a "clean" ASCII string to pass to the native
API.

Using URI's for input slaves is new, but the functionality
regarding URI's is common with that described above for MRL's - so
this change factors the common code into a new NativeUri utility
class.

This change also cleans up the sub-title file API in the base
media player implementation class.