Remove logging from the bespoke Linux initialisation
Rationale is firstly, errors simply should not happen due to the
type of initialisation code in this class. Secondly, even if errors
do occur, we can not do anything about them. This is something that
at run-time either works or doesn't work.

If there are errors, we robustly ignore them and try and continue
regardless.