Refactoring of direct media player (video)
The "direct rendering" media player now just uses a custom video
surface implementation for the EmbeddedMediaPlayer rather than
requiring a whole specialised component of its own.

Everything previously in the direct rendering media player is
reimplemented as a video surface and a helper class. The existing
direct media player is gone entirely.

The DirectMediaPlayerComponent is also  gone and replaced with a
specialised EmbeddedMediaPlayerComponent instead - the new
CallbackMediaPlayerComponent and the existing embedded version
both extend EmbeddedMediaPlayerComponentBase.

The new callback media player component provides a reasonable
default implementation for rendering the video, but as expected
a client application can override the defaults and provide their
own implementation of a RenderCallback as was the case with old
direct media player.

Tests were also updated to use the new callback media player and
component.

This is all mostly working, with the notable exception of pending
changes to the type-safe component builders and the component spec
builders which will follow.