Changes to type-safe builders and component specs
Component specs are kinda-builders that are used to build the
constructor parameters only, not build the actualy component.

Move the spec classes from the specific component base classes
to a consolidated MediaPlayerSpecs class.

Change the component builder to support building the new
CallbackMediaPlayerComponent.

The component builder is currently questionable, as it now allows
you to set parameter combinations that are invalid - for example
if you specify an overlay, you should not be able to specify a
callback media player. The fluent API used currently is a naive
in this regard.

Since we have the component spec builders, perhaps the component
builder itself is extraneous and we can side-step this issue by
deleting it.

In fact, the specs could optionally create a new instance of the
class being spec'd, which gives us a reasonable compromise
solution for this issue.