Completely internalise the current Media instance
A lot of difficulties arise from who exactly owns the Media
instance and consequently who is responsible for its lifecycle,
especially with regard to releasing that Media instance.

As far as the media player is concerned, the media player itself is
now entirely responsible for an internal Media instance that will
be created each time new media is prepared/played/started by one or
other of those methods.

A client application still has access to those services/aspects of
the media instance by a methods on the media player MediaService
that delegate to the internal media instance.

With this approach there is no ambiguity - the media player itself
creates the Media instance, releases it at the appropriate time,
and client applications are prevented from interfering with this
lifecycle by API restrictions (i.e. there is simply no API provided
to "mess" with the internal media instance).

These changes break one particular test that requires a Media
instance which is now no longer available. This will be addressed
in a subsequent change.