Complete internalise the current MediaList instance
This is similar to the previous changes around the Media instance
in the base media player.

To prevent any ambiguity, the MediaList instance is now fully
"owned" by the component itself and is not directly exposed to
client applications.

Applications can still access the various aspects of the media list
by using delegated functions.