Consolidate asynchronous task executors
This is intended to be used by event handling code that may wish
to call back into LibVLC as a result of a native event. Generally
it is not allowed to callback, strange behaviour at best or a fatal
JVM crash at worse can happen when calling back from the native
event thread.

So, event handlers can submit a task that will be off-loaded from
the native callback thread and executed asynchronously.

Factored the duplicate implementation in MediaPlayer and
MediaListPlayer to use a new common component.

Added an equivalent implementation to the MediaPlayerFactory for
those applications that are not using a MediaPlayer or
MediaListPlayer or for more general usage in any case.

Updated Javadoc.