Have the media player manage hold/release on the renderer item
Should make client applications a bit simpler.

The implementation is a little more intricate than may be thought.

This method holds the renderer (i.e. invokes hold), if it is not
null, before setting it. If setting the renderer subsequently fails
this method will invoke release on the renderer before returning
the error.

If setting the new renderer fails, the old one is kept.