Add new video surface that gets the component id as needed
Using this video surface it is possible to set the video surface
for the media player in advance (as usual), but also to defer when
the component id is obtained. So, rather than providing the
component id up front, it can be provided later when the media
player actually needs it.

The component id in this implementation is provided by sub-classes
implementing a template method.