Remove the synchronisation wrapper on the JNA native library
This was causing real deadlocks.

This was initially hesitantly added as a defensive safety measure,
although it was never used in any vlcj version right up to 4.0.0.

It was intended to prevent concurrent access into LibVLC as it was
thought this was a Bad Thing, but even that assumption is not
clear.