Fix the callback video surface image painters
Do not use g2.setTransform(), instead use g2.translate() and
g2.scale().

This is because the graphics context may already have a transform
set on it which would be clobbered by setTransform().

This was apparent when the video surface had a menu bar, opening
the menu would change the drawing offset.