Simplify callback video surface image painters
Seems simpler to just recalculate the scale factors and offsets
every frame rather than do conditional checks and stores/reads to
and from class fields.

This also has the benefit of resolving an issue when changing
media whereby the painter was painting at the resolution of the
previous video until the window was resized.