Add a note on media player stopped events
It is possible that stopped events may be received even if the media
player is not currently playing.

This can happen when changing media as it may trigger the media list
player that is associated with the media player to explicitly invoke
stop. This explicit invocation is required for proper sub-item
handling when dealing with media that contains sub-items.

What this means in practical terms is that client applications must
be prepared for these potentially spurious stopped events to occur
when perhaps they are not expected to. Ignoring the event would be
a reasonable strategy.