Add callback method to notify applications of the native buffers
There are use-cases where it will be very handy for the calling
application to know the native video buffers that were allocated so
that the buffer(s) can be shared e.g. by an OpenGL texture or the
proposed PixelBuffer class in JavaFX. Sharing the buffer reduces
the need for full-frame buffer copies.