Change media player event manager to suppress spurious stopped events
vlcj 4.x introduced behaviour (related to media with sub-items, like
YouTube media) whereby an associated media list player was always
explicitly stopped when changing media. An unintended side-effect of
this was for a spurious media player "stopped" event to be raised
before the media was played.

Getting a stopped event before a played event is undesirable and
confusing behaviour.

This commit adds a mechanism to suppress the generation of this
stopped event - now, each time media is changed, a stopped event will
be suppressed unless there was a played event generated prior. The
mechanism will be reset each time the media is changed.

This should have no effect where media is played/stopped/played etc.