Changes to full-screen for macOS for Java9+
The EAWT classes used to enter native full screen are no longer
available if using the Java Module System.

There is no working alernative, so full screen on macOS is itself
no longer available.

This will require a change to the JDK to work.

The idea of this change is to allow vlcj to be used on macOS on
Java9+, by using reflection for earlier than Java 8 (where full
screen will still work) rather than linking to the EAWT classes.