Move the bespoke Linux initialisation to the video surface
Previously this initialisation was in the MediaPlayerFactory and it
was executed whether it was needed or not (e.g. it is not needed in
a JavaFX application).

In fact, executing this initialisation in a JavaFX environment led
to fatal JVM crashses in some circumstances (like opening a native
file chooser dialog).

The solution therefore is to create a new base class for video
surfaces that use AWT and put the initialisation in there.

This change also removes the need to pass the VLCJ_INITX=no system
property that was previously used to workaround this issue.