Native discovery can only be performed once
Change the boolean to a static so there can be multiple factory
instances, but the native discovery will only run once in total
rather than once per factory.

Once the native discovery has completed successfully and loaded the
native library, it is impossible to load that native library from
anywhere else and so it is consequently pointless to run the native
discovery again.