Support rendering video with non 1:1 SAR in the callback media player
This is used e.g. for DVD video where the pixels are not square (1:1).

Main changes are to notify the image painters when a new video track
is selected and to provide a new implementation of an image painter
that knows how to use SAR.

This change was only feasible because of the new native track
selection API that comes with LibVLC 4.0.0.