Changes to video engine (OpenGL) rendering
The update-output callback is NOT used for setting the width/height
rather it reports the width/height that was previously set.

The width/height is supposed by set by invoking the new native
report-size-changed callback.

For update-output the idea is that the native rendering configuration
context(s) can be adjusted by the client. This is an inherently low-
level operation, and at the moment the native LibVLC structures are
exposed on this callback interface. This may change in the future,
but for now there is little point wrapping that structure in a non-
native object.