Add a custom timer event and listener
The idea is to provide a smoother (more regular) media playback
timer.

Just relying on the native media_player_time_changed event can
produce erratic timer notifications, so a simple clock display might
seem to have long or short seconds.

Instead, use the native libvlc_clock.

With this implementation there may be circumstances where the media
is paused or stopped and there is a pending timer notification - in
such cases that pending notification will NOT (currently) be sent.

This may be revisited.