Use actual video track size when rendering
Previously it used the full video buffer size, and it is possible
that the video buffer is larger than the actual video image (e.g.
because the video buffer is aligned in memory to a specific byte
boundary).

Without this, the video scaling would be slightly off when rendering
and lead to unnecessary black bars.

Note that it is possible a few frames may be rendered before the
video track information is known.