Add convenience constructors to pass LibVLC initialisation args
This is a common use-case, i.e. to use the default configurations
of the component classes but with specific LibVLC args.

For the audio components there is no point as they only accept a
MediaPlayerFactory anyway (at least at the present time) - and
adding an equivalent constructor here can lead to an ambigious
constructor signature.