Explicitly load libvlccore on all platforms
The java.library.path system property is only used when loading a
library explicitly with System.loadLibrary().

Any dependent libraries loaded by that library will NOT be found
via java.library.path - they will instead be located by the runtime
OS library loading method, and this will likely fail.

It does fail on Windows 10 and macOS, although it seems to be OK
on Linux.

It is however harmless to do this in Linux as well, so simplify
matters by loading libvlccore on all platforms.