Restore "finished" event behaviour
This is a (potentially) temporary change.

With LibVLC 4.x (currently) the semantics of stopped and finished
events has changed as compared to LibVLC 3.x.

The "finished" event was changed to "stopping", and as of the date
of this commit it does not guarantee that the media reached its
end.

There is an outstanding issue with VLC on this topic.