Provide display width/size during buffer format callback
With LibVLC 3.x only the video buffer size is passed via the
callback, and this can provide unexpected values like 1090 instead
of 1080 for height.

With LibVLC 4.x, both the video buffer size and the display size
are passed.

For future compatibility with vlcj 5.x, this behaviour can be
emulated by using the video dimension API to get the display size.