Partially revert the callback display size changes
These changes were made to harmonise the API with upcoming vlcj 5
and LibVLC 4 changes.

However, for certain videos, the heuristic approximation of the
LibVLC 4 changes caused invalid video dimensions.

For example, the API would report a width of 1408 for the buffer,
but a display size would be reported as 1400. This discrepancy
caused garbling of the video output.

So for vlcj 4 and LibVLC 3, this was undone to just use the size
reported by LibVLC.