Remove the UserDir discovery provider in favour of new AppDir provider
The intention for the user dir provider was always to return the
same directory that the application was installed in - in almost
all cases this would be expected to be the "current" directory as
obtained from the user.dir system property.

However, the current directory is not necessarily the same as the
installation directory.

So instead, remove that provider and replace it with one that uses
class loader protection domain and code source to find the file
location containing the vlcj class files and return this as the
application directory.